// client/ts/src/manifest/types/ClaimedSeat.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as web3 from '@solana/web3.js';
import * as beet from '@metaplex-foundation/beet';
import * as beetSolana from '@metaplex-foundation/beet-solana';
export type ClaimedSeat = {
  trader: web3.PublicKey;
  baseWithdrawableBalance: beet.bignum;
  quoteWithdrawableBalance: beet.bignum;
  quoteVolume: beet.bignum;
  padding: number[] /* size: 8 */;
};

/**
 * @category userTypes
 * @category generated
 */
export const claimedSeatBeet = new beet.BeetArgsStruct<ClaimedSeat>(
  [
    ['trader', beetSolana.publicKey],
    ['baseWithdrawableBalance', beet.u64],
    ['quoteWithdrawableBalance', beet.u64],
    ['quoteVolume', beet.u64],
    ['padding', beet.uniformFixedSizeArray(beet.u8, 8)],
  ],
  'ClaimedSeat',
);