// client/ts/src/manifest/types/RestingOrder.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet';
import { OrderType, orderTypeBeet } from './OrderType';
export type RestingOrder = {
  price: beet.bignum;
  numBaseAtoms: beet.bignum;
  sequenceNumber: beet.bignum;
  traderIndex: number;
  lastValidSlot: number;
  isBid: boolean;
  orderType: OrderType;
  padding: number[] /* size: 22 */;
};

/**
 * @category userTypes
 * @category generated
 */
export const restingOrderBeet = new beet.BeetArgsStruct<RestingOrder>(
  [
    ['price', beet.u128],
    ['numBaseAtoms', beet.u64],
    ['sequenceNumber', beet.u64],
    ['traderIndex', beet.u32],
    ['lastValidSlot', beet.u32],
    ['isBid', beet.bool],
    ['orderType', orderTypeBeet],
    ['padding', beet.uniformFixedSizeArray(beet.u8, 22)],
  ],
  'RestingOrder',
);